+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 14.0.0
+* upgrade to webpack 4
+* last version to support gulp 3.x
+
+## 12.0.0
+* Dependency upgrade, considered a breaking change due to eslint rule changes
+
+## 11.0.0
+* Upgrade to webpack 3
+
+## 10.0.0
+* Dependency upgrade, considered a breaking change due to eslint rule changes
+
 ## 9.0.0
 * Can now set a custom .eslintrc and tsconfig location
 * update to airbnb 14.1.0